Rename logging.py to log.py to avoid headaches with name collision with logging package
